+++ lib/larva/configuration.rb
@@ -3,6 +3,10 @@ module Larva
   end
 
   class Configuration
+    attr_accessor :logger
 
+    def initialize
+      @logger = Filum.logger
+    end
   end
 end